UBUNTU: SAUCE: dell-laptop: fire SMI when toggling hardware killswitch (revised)
BugLink: http://bugs.launchpad.net/bugs/590607

SRU Justification:

Impact:
A type of Dell Inspiron Mini does not turn off bluetooth physically when
pressing F2 (rfkill hotkey).

Fix:
Current kernel code expects hardware to handle the hard-rfkill switching
spontaneously and only notify rfkill subsystem with the status change.
This patch makes kernel to explicitly fire a SMI to switch on/off
rfkill devices if BIOS reports that the hardware switch is not supported.

This patch depends on the following sauce patch in lucid kernel so is not going upstream:
        UBUNTU: SAUCE: dell-laptop: Store the HW switch status internally rather than requerying every time

Testcase:
I tested the patch on a Dell Mini 10. It works smoothly.

Signed-off-by: Keng-Yu Lin <keng-yu.lin@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Steve Conklin <sconklin@canonical.com>